requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-pyc==1.1.0
django-rest-authtoken==2.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
python==3.9.1
sqlparse==0.4.1
